DB/Quests: Restore pre-cataclysm spanish translations.
Added updates to translations, in the try to remove cataclysm changes that got merged in the translations (only the spanish ones).
At least, this is an attempt to not be confused anymore for the changed quests, changed locations, and so, as many of the quests changed also the finish location it tells under its
objectives.
See notes in the beginning of the archive.
Closes #15076

Signed-off-by: jackpoz <giacomopoz@gmail.com>