When crafting a item cap recipe ilevel if set
When crafting a item through a spell effect, the item level is capped at the value set in the item_template db entry. If the item is being regenerated at a fixed ilevel then this condition is ignored. Dungeon loot remains uncapped because it is modified by the groups average ilevel.